Xcode 5.1 Compatibility Fixes

+++ framework/core/oniguruma/regint.h
@@ -279,7 +279,7 @@ typedef unsigned int  BitStatusType;
 
 #define INT_MAX_LIMIT           ((1UL << (SIZEOF_INT * 8 - 1)) - 1)
 
-#define DIGITVAL(code)    ((code) - '0')
+#define DIGITVAL(code)    ((int) ((code) - '0'))
 #define ODIGITVAL(code)   DIGITVAL(code)
 #define XDIGITVAL(enc,code) \
   (ONIGENC_IS_CODE_DIGIT(enc,code) ? DIGITVAL(code) \
@@ -575,7 +575,7 @@ typedef void* PointerType;
 #define GET_STATE_CHECK_NUM_INC(num,p)  PLATFORM_GET_INC(num, p, StateCheckNumType)
 
 /* code point's address must be aligned address. */
-#define GET_CODE_POINT(code,p)   code = *((OnigCodePoint* )(p))
+#define GET_CODE_POINT(code,p)   code = (int) *((OnigCodePoint* )(p))
 #define GET_BYTE_INC(byte,p) do{\
   byte = *(p);\
   (p)++;\
